move layout components types to component page folder
