+++ components/layouts/Footer/__tests__/footer.spec.tsx
@@ -3,7 +3,7 @@ import { Provider } from 'react-redux';
 import configureStore from 'redux-mock-store';
 import { render } from '@testing-library/react';
 
-import reducer from '../../../../store/reducer';
+import reducer from '../../../../store/redux/reducer';
 
 import { Footer } from '../Footer';
