+++ components/pages/DetailPage/RelatedRestaurants.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { Card } from '../../core/Card/Card';
 
-import { Restaurant } from '../../../helpers/types';
+import { Restaurant } from '../types';
 
 import { THUMB_GENERIC_SRC } from '../../../store/statics';
