+++ components/pages/HomePage/HomePage.tsx
@@ -8,7 +8,7 @@ import { Card } from '../../core/Card/Card';
 import { StyledHighlights, StyledHighlightWrapper } from './styles';
 
 import { CardType } from '../../core/types';
-import { HighlightRestaurant } from '../../../helpers/types';
+import { HighlightRestaurant } from '../types';
 
 type HomePageProps = {
   highlights: HighlightRestaurant[];
