+++ components/ui/Finder/__tests__/finder.spec.tsx
@@ -4,7 +4,7 @@ import configureStore from 'redux-mock-store';
 import { render } from '@testing-library/react';
 import 'jest-styled-components';
 
-import reducer from '../../../../store/reducer';
+import reducer from '../../../../store/redux/reducer';
 
 import { Finder } from '../Finder';
