+++ helpers/__tests__/utils.spec.tsx
@@ -15,7 +15,8 @@ import {
   FOURTH_DETAIL_MOCKED,
   RELATED_RESTAURANTS_MOCKED,
 } from '../../components/pages/DetailPage/__mocks__/detailpage.mocks';
-import { LOCATIONS } from '../staticData';
+
+import { LOCATIONS } from '../../store/statics';
 
 describe('compose', () => {
   it('should be equal to mocked text', () => {
