+++ helpers/utils.tsx
@@ -1,4 +1,4 @@
-import { Restaurant } from './types';
+import { Restaurant } from '@components/pages/types';
 
 // GENERAL UTILS
