+++ pages/index.tsx
@@ -1,27 +1,23 @@
 import React from 'react';
-
 import { NextPage, InferGetStaticPropsType } from 'next';
 import { useRouter } from 'next/router';
 import dynamic from 'next/dynamic';
-
 import { useDispatch } from 'react-redux';
 
 import { FullLoader } from '@components/ui/FullLoader/FullLoader';
-
 import { Loader } from '@components/core/Loader/Loader';
-
 import { useBreadcrumbs } from '@components/hooks/useBreadcrumbs';
 
 import { setRelatedRestaurants } from '@store/redux/actions';
-
-import { HighlightRestaurant } from '@helpers/types';
-
 import {
   DEFAULT_TEXT_LOADING,
   DEFAULT_BREADCRUMB,
   HIGHLIGHTED_RESTAURANTS,
 } from '@store/statics';
+
 import { getCurrentRelatedRestaurants } from '@helpers/utils';
+
+import { HighlightRestaurant } from '@components/pages/types';
 
 type HomeProps = InferGetStaticPropsType<typeof getStaticProps>;
